+++ package.json
@@ -38,6 +38,7 @@
         "rollup-plugin-copy": "^3.5.0",
         "rollup-plugin-postcss": "^4.0.2",
         "rollup-plugin-serve": "^2.0.2",
+        "sass": "^1.69.4",
         "typescript": "^5.2.2",
         "vitest": "^0.34.6"
     },
