+++ src/components/Carousel.scss
