+++ src/components/Heading.js
@@ -5,7 +5,7 @@
 import { h1, h2, h3, h4, h5, h6 } from '../libs/makeElement';
 import { cx } from '../utils/dom/cx';
 
-import s from './Heading.css';
+import s from './Heading.scss';
 
 /**
  *
