+++ src/components/Keyboard.scss
