+++ src/components/LazyImage.scss
