+++ src/components/Notification.scss
