+++ src/components/Tile.scss
