+++ src/declarations/assets.d.ts
@@ -2,3 +2,4 @@ declare module '*.png';
 declare module '*.jpg';
 declare module '*.svg';
 declare module '*.css';
+declare module '*.scss';
