+++ src/styles/_colours.scss
@@ -1,0 +1,48 @@
+// Primary
+$primary: #FFDE2B;
+$primaryXLight: #FEEB86;
+$primaryLight: #BAA21F;
+$primaryDark: #806F15;
+$primaryDarker: #474432;
+$primaryDisabled: #857730;
+
+// Success
+$success: #02C07D;
+$successLight: #7EDEBC;
+$successDark: #245D49;
+
+// Warning
+$warning: #FF6700;
+$warningLight: #FFE1BE;
+$warningDark: #401A00;
+
+// Error
+$error: #DE1C22;
+$errorLight: #F9D3D4;
+$errorDark: #300006;
+
+// Info
+$info: #2196F3;
+$infoLight: #60B2F3;
+$infoDark: #2E516D;
+
+// Delete
+$delete: #E84545;
+$deleteLight: #EC6C6C;
+$deleteDark: #AD7676;
+
+// Gray
+$black: #000;
+$onyx: #1E1E1E;
+$cavern: #242424;
+$slate: #2C2C2C;
+$pewter: #333;
+$shadow: #373737;
+$charcoal: #3B3B3B;
+$graphite: #4C4C4C;
+$mist: #616161;
+$dust: #8A8A8A;
+$neutral: #A8A8A8;
+$marshmellow: #D0D0D0;
+$bone: #E0E0E0;
+$white: #FFF;
