+++ src/views/canivideo/canivideo.scss
