+++ src/views/search.scss
