+++ src/views/show.scss
